Consolidate ItemView from static layers
